Changed image header to use uint16_t for width and height.

+++ inc/img.h
@@ -25,8 +25,8 @@ typedef enum {
 /* Image structure */
 typedef struct {
   /* Image size & depth */
-  uint8_t width;
-  uint8_t height;
+  uint16_t width;
+  uint16_t height;
   uint8_t depth;
 
   /* Image type. */
